pyproject.toml
[tool.poetry]
name = "documentation-helper"
version = "0.1.0"
description = ""
authors = ["Igorithm <amit1994biswas@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
langchain = "^0.2.6"
openai = "^1.35.7"
streamlit-chat = "^0.1.1"
langchain-community = "^0.2.6"
python-dotenv = "^1.0.1"
langchain-pinecone = "^0.1.1"
langchain-openai = "^0.1.13"
langchain-text-splitters = "^0.2.2"
unstructured = "^0.14.9"
psutil = "^6.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
